+++ packages/contracts/contracts/Handler.sol
@@ -41,6 +41,13 @@ contract Handler is IHandler, BalanceManager, OwnableUpgradeable {
         _;
     }
 
+    function addToAssetPrefill(
+        EncodedAsset calldata encodedAsset,
+        uint256 value
+    ) external onlyOwner {
+        _addToAssetPrefill(encodedAsset, value);
+    }
+
     // Gives an address permission to call `fillBatchesWithZeros`
     function setSubtreeBatchFillerPermission(
         address filler,